+++ src/js/components/Overview/PublicOverview.tsx
@@ -18,7 +18,7 @@ import { useAppSelector } from '@/hooks';
 import { useSelectedProject, useSelectedScope } from '@/features/metadata/hooks';
 import { useTranslation } from 'react-i18next';
 import { RequestStatus } from '@/types/requests';
-import Catalogue from '@/components/Provenance/Catalogue';
+import Catalogue from '@/components/Provenance/Catalogue/Catalogue';
 
 const ABOUT_CARD_STYLE = { width: '100%', maxWidth: '1390px', borderRadius: '11pX', ...BOX_SHADOW };
 const MANAGE_CHARTS_BUTTON_STYLE = { right: '5em', bottom: '1.5em', transform: 'scale(125%)' };
