+++ src/js/components/Provenance/Catalogue/CatalogueCard.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Card, Carousel, Descriptions, Flex, Space, Tag, Typography } from 'antd';
-import CatalogueCarouselDataset from '@/components/Provenance/CatalogueCarouselDataset';
+import CatalogueCarouselDataset from '@/components/Provenance/Catalogue/CatalogueCarouselDataset';
 import type { Project } from '@/types/metadata';
 import { ISODateToString } from '@/utils/strings';
